chore(Enhancement) Merge step and action class implementation
