+++ keep/alert/alert.py
@@ -4,7 +4,6 @@ import typing
 
 from pydantic.dataclasses import dataclass
 
-from keep.action.action import Action
 from keep.contextmanager.contextmanager import ContextManager
 from keep.iohandler.iohandler import IOHandler
 from keep.step.step import Step, StepError
@@ -22,9 +21,9 @@ class Alert:
     alert_owners: typing.List[str]
     alert_tags: typing.List[str]
     alert_steps: typing.List[Step]
-    alert_actions: typing.List[Action]
+    alert_actions: typing.List[Step]
     alert_file: str = None
-    on_failure: Action = None
+    on_failure: Step = None
 
     def __post_init__(self):
         self.logger = logging.getLogger(__name__)
@@ -68,7 +67,7 @@ class Alert:
                 raise
         self.logger.debug(f"Steps for alert {self.alert_id} ran successfully")
 
-    def run_action(self, action: Action):
+    def run_action(self, action: Step):
         self.logger.info("Running action %s", action.name)
         try:
             action_status = action.run()
